moved timeline queries to session queries
